% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_cross.R
\name{chr_cols}
\alias{chr_cols}
\title{Identify character columns in a data frame}
\usage{
chr_cols(df, cols = NULL)
}
\arguments{
\item{df}{A dataframe to check for character columns.}

\item{cols}{An optional vector of column names to check.
If NULL (default), all columns in the dataframe are checked.}
}
\value{
A character vector containing the names of the columns
in the dataframe that are of character type.
}
\description{
This function identifies which columns in a data frame are of character type.
}
\examples{
chr_cols(dummy_data)

}
