% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{Dummy data for enquete.}
\format{
A data frame with 50 rows and 4 variable:
\describe{
\item{q1}{Answers of single choice for categorical scales.}
\item{q2_ord}{Answers of single choice for ordinal scales.}
\item{q3_multi}{Answers of multiple choice for categorical scales.}
\item{q4_comment}{Answers of fee descriptive comments.}
}
}
\usage{
dummy_data
}
\description{
Dummy data for enquete.
}
\examples{
data(dummy_data)
dummy_data
}
\keyword{datasets}
