% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_simple.R
\name{enq_simple}
\alias{enq_simple}
\title{Count occurrences of values in a column}
\usage{
enq_simple(df, col, delim = ";")
}
\arguments{
\item{df}{A dataframe containing the column to be counted.}

\item{col}{A string for column to be count.}

\item{delim}{A string to be used for delimiter of values
in the column (default is ";").}
}
\value{
A dataframe with two columns:
"option" (the unique values after splitting) and "n"
(the number of occurrences of each unique value).
}
\description{
This function splits values in a character column by a delimiter
and then counts the occurrences of each unique value.
It uses \code{tidyr::separate_longer_delim()} to split the column and
\code{dplyr::count()} to count the occurrences. The order of the output
is determined by the presence of "_ord" in the column name.
If "_ord" is present, the output is arranged by "option",
otherwise it is arranged by "n" (number of occurrences).
}
\examples{
enq_simple(dummy_data, "q1")
enq_simple(dummy_data, "q2_ord")
enq_simple(dummy_data, "q3_multi")

}
