% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_summary.R
\name{enq_summary}
\alias{enq_summary}
\alias{enq_summary_bind}
\title{Create summaries for multiple character columns in a data frame}
\usage{
enq_summary(df, cols_except = "_comment$", delim = ";", simplify = FALSE)

enq_summary_bind(df)
}
\arguments{
\item{df}{A dataframe containing the column to be counted.}

\item{cols_except}{A regular expression pattern to exclude columns from the summary.
Columns that do not match the pattern will be included.
Defaults to "_comment$".}

\item{delim}{A string to be used for delimiter of values
in the column (default is ";").}

\item{simplify}{A logical. If TRUE, bind elements of a list (default is FALSE).}
}
\value{
A list of dataframe.
Names of the list elements are the same as the column names.
}
\description{
This function creates summaries for
multiple character columns in a data frame.
It uses \code{enq_simple} to calculate the occurrences of
split values for each column.
}
\examples{
enq_summary(dummy_data)
enq_summary(dummy_data, simplify = TRUE)

}
