% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensumr-package.R
\docType{package}
\name{ensumr-package}
\alias{ensumr}
\alias{ensumr-package}
\title{ensumr: Analyse Enquate Data Easily}
\description{
Analyse Enquate Data Easily.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matutosi/ensumr}
  \item \url{https://matutosi.github.io/ensumr/}
}

}
\author{
\strong{Maintainer}: Toshikazu Matsumura \email{matutosi@gmail.com} (\href{https://orcid.org/0000-0003-2306-6355}{ORCID})

}
\keyword{internal}
