% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enq_cross.R
\name{split_by}
\alias{split_by}
\title{Shortcut for \code{base::split()}}
\usage{
split_by(df, group)
}
\arguments{
\item{df}{A dataframe.}

\item{group}{A string.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A list of split dataframe.
}\if{html}{\out{</div>}}
}
\description{
Shortcut for \code{base::split()}
}
